added code styles
