--- src/components/banner-main-page/index.ts
@@ -1,1 +1,0 @@
-export { BannerMainPage } from './BannerMainPage';
