+++ src/components/banner-on-main-page/BannerOnMainPage.module.scss
