+++ src/components/banner-on-main-page/index.ts
@@ -1,0 +1,1 @@
+export { BannerOnMainPage } from './BannerOnMainPage';
