+++ src/components/banner-on-main-page/props.ts
